Move classes into a hirearchy of use_cases/helpers/model
This helps to differentiate responsibilities for each class.
