+++ lib/model/user.rb
