+++ lib/use_cases/email_sponsees_extractor.rb
