+++ spec/lib/model/user_spec.rb
