+++ spec/lib/use_cases/email_signup_spec.rb
