+++ spec/lib/use_cases/email_sponsees_extractor_spec.rb
