+++ spec/lib/use_cases/sms_template_finder_spec.rb
